Implement support for split single courses
These ones are actually a lot more common than the multi-courses.

Since they're grades are combined in the end by a weighted average,
there will actually always be the complete grade for the course too.

So a course split into two parts will be reported as one main course and
two subcourses.

This also changes the loading behavior.
With these split courses it is pretty common that parts are only
actually added to the tables when their grades are published.
Therefore new courses that pop up and are already scored will now be
detected as updates.